requirements.txt
# Core dependencies
websockets>=12.0

# macOS (uses pyusb for direct USB access)
pyusb>=1.2.1; sys_platform == 'darwin'
rumps>=0.4.0; sys_platform == 'darwin'

# Windows (uses pyserial for COM port - works with Sartorius Driver PMA)
pyserial>=3.5; sys_platform == 'win32'
pystray>=0.19.0; sys_platform == 'win32'
pillow>=10.0.0; sys_platform == 'win32'

# Build tools
pyinstaller>=6.0.0
